Split up the tests and actually enqueue the events from the WebhookProcessor instead of doing them inline
